feat: move email capture to root model
